Use asyncstorage-mock-another to make tests run headless without React Native

+++ test/testCommon.js
@@ -11,7 +11,7 @@ var lastLocation = function () {
 
 var cleanup = function (callback) {
 
-  if (window.localStorage) {
+  if (typeof window !== 'undefined' && window.localStorage) {
     window.localStorage.clear();
   }
